Remove nvcuvid and nvencodeapi
nvcudvid and nvencodeapi does not really see much use, and there
are probably just aswell to leave it since no change is done over
the staging implementation IF it should be needed.

Afaik there is no Steam related apps/games that run under Proton
that actually needs these, so it is moved to the "old" branch
for now.